Manage Task and Status Update Done

+++ backend/src/main/java/org/management/backend/controller/UserController.java
@@ -26,7 +26,7 @@ public class UserController {
     }
 
     @PutMapping("/{id}")
-    public User updateUser(@PathVariable int id, @RequestBody User userDetails) {
+    public User updateUser(@PathVariable Long id, @RequestBody User userDetails) {
         User user = userRepository.findById(id).orElseThrow();
         user.setName(userDetails.getName());
         user.setEmail(userDetails.getEmail());
@@ -35,7 +35,7 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable int id) {
+    public void deleteUser(@PathVariable Long id) {
         userRepository.deleteById(id);
     }
 }
